Split agent action parsing helpers
