+++ Sources/QuillCodeAgent/AgentActionJSONParser.swift
@@ -5,9 +5,11 @@ import QuillCodeTools
 
 public enum AgentActionJSONParser {
     public static func parse(_ text: String) throws -> AgentAction {
-        let trimmed = stripFences(text.trimmingCharacters(in: .whitespacesAndNewlines))
-        guard let object = actionObject(in: trimmed) else {
-            if let recovered = recoveredShellAction(from: trimmed) {
+        let trimmed = AgentActionJSONExtractor.strippedFences(
+            from: text.trimmingCharacters(in: .whitespacesAndNewlines)
+        )
+        guard let object = AgentActionJSONExtractor.actionObject(in: trimmed, looksLikeAction: looksLikeActionObject) else {
+            if let recovered = AgentShellCommandRecovery.recoveredAction(from: trimmed) {
                 return recovered
             }
             throw TrustedRouterAgentError.invalidActionJSON(text)
@@ -26,7 +28,7 @@ public enum AgentActionJSONParser {
             var arguments = canonicalArguments(for: name, in: object)
             if name == ToolDefinition.shellRun.name,
                arguments["cmd"] == nil,
-               let recoveredCommand = recoverExplicitShellCommand(from: trimmed) {
+               let recoveredCommand = AgentShellCommandRecovery.explicitCommand(from: trimmed) {
                 arguments["cmd"] = recoveredCommand
             }
             if arguments.isEmpty && Self.requiresNonEmptyArguments(name) {
@@ -45,170 +47,8 @@ public enum AgentActionJSONParser {
         }
     }
 
-    private static func actionObject(in text: String) -> [String: Any]? {
-        if let object = parseObject(text), looksLikeActionObject(object) {
-            return object
-        }
-        for candidate in jsonObjectCandidates(in: text) {
-            guard let object = parseObject(candidate), looksLikeActionObject(object) else { continue }
-            return object
-        }
-        return nil
-    }
-
-    private static func parseObject(_ text: String) -> [String: Any]? {
-        guard let data = text.data(using: .utf8) else { return nil }
-        return (try? JSONSerialization.jsonObject(with: data)) as? [String: Any]
-    }
-
     private static func looksLikeActionObject(_ object: [String: Any]) -> Bool {
         object["type"] is String || toolName(in: object) != nil
-    }
-
-    private static func jsonObjectCandidates(in text: String) -> [String] {
-        var candidates: [String] = []
-        var startIndex: String.Index?
-        var depth = 0
-        var isInsideString = false
-        var isEscaping = false
-
-        for index in text.indices {
-            let character = text[index]
-            guard let start = startIndex else {
-                if character == "{" {
-                    startIndex = index
-                    depth = 1
-                    isInsideString = false
-                    isEscaping = false
-                }
-                continue
-            }
-
-            if isInsideString {
-                if isEscaping {
-                    isEscaping = false
-                } else if character == "\\" {
-                    isEscaping = true
-                } else if character == "\"" {
-                    isInsideString = false
-                }
-                continue
-            }
-
-            if character == "\"" {
-                isInsideString = true
-            } else if character == "{" {
-                depth += 1
-            } else if character == "}" {
-                depth -= 1
-                if depth == 0 {
-                    candidates.append(String(text[start...index]))
-                    startIndex = nil
-                }
-            }
-        }
-
-        return candidates
-    }
-
-    private static func recoveredShellAction(from text: String) -> AgentAction? {
-        guard let command = recoverExplicitShellCommand(from: text) else {
-            return nil
-        }
-        return .tool(.init(
-            name: ToolDefinition.shellRun.name,
-            argumentsJSON: ToolArguments.json(["cmd": command])
-        ))
-    }
-
-    private static func recoverExplicitShellCommand(from text: String) -> String? {
-        let spans = inlineCodeSpans(in: text)
-        for span in spans {
-            let command = span.code.trimmingCharacters(in: .whitespacesAndNewlines)
-            guard isPlausibleShellCommand(command),
-                  hasExecutionIntent(before: span.range.lowerBound, in: text)
-            else {
-                continue
-            }
-            return command
-        }
-        return nil
-    }
-
-    private static func inlineCodeSpans(in text: String) -> [(code: String, range: Range<String.Index>)] {
-        var spans: [(String, Range<String.Index>)] = []
-        var searchIndex = text.startIndex
-        while searchIndex < text.endIndex,
-              let opening = text[searchIndex...].firstIndex(of: "`") {
-            let afterOpening = text.index(after: opening)
-            guard afterOpening < text.endIndex else { break }
-            if text[afterOpening] == "`" {
-                searchIndex = afterOpening
-                continue
-            }
-            guard let closing = text[afterOpening...].firstIndex(of: "`") else { break }
-            if text[text.index(before: closing)] != "`" {
-                spans.append((String(text[afterOpening..<closing]), opening..<text.index(after: closing)))
-            }
-            searchIndex = text.index(after: closing)
-        }
-        return spans
-    }
-
-    private static func hasExecutionIntent(before index: String.Index, in text: String) -> Bool {
-        let lowerBound = text.index(index, offsetBy: -96, limitedBy: text.startIndex) ?? text.startIndex
-        let prefix = text[lowerBound..<index]
-            .lowercased()
-            .replacingOccurrences(of: "\n", with: " ")
-        let negativeIntents = [
-            "do not run",
-            "don't run",
-            "will not run",
-            "won't run",
-            "cannot run",
-            "can't run",
-            "should not run",
-            "do not execute",
-            "don't execute",
-            "will not execute",
-            "won't execute"
-        ]
-        guard !negativeIntents.contains(where: { prefix.contains($0) }) else {
-            return false
-        }
-        let intents = [
-            "i'll run",
-            "i’ll run",
-            "i will run",
-            "i'll execute",
-            "i’ll execute",
-            "i will execute",
-            "i'll check",
-            "i’ll check",
-            "i will check",
-            "i am running",
-            "i'm running",
-            "i’m running",
-            "running",
-            "run ",
-            "execute "
-        ]
-        return intents.contains { prefix.contains($0) }
-    }
-
-    private static func isPlausibleShellCommand(_ command: String) -> Bool {
-        let trimmed = command.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !trimmed.isEmpty,
-              !trimmed.contains("\n"),
-              trimmed.count <= 500
-        else {
-            return false
-        }
-        let firstWord = trimmed.split(separator: " ", maxSplits: 1).first.map(String.init) ?? trimmed
-        guard firstWord.range(of: #"^[A-Za-z0-9_./:-]+$"#, options: .regularExpression) != nil else {
-            return false
-        }
-        return true
     }
 
     private static func toolName(in object: [String: Any]) -> String? {
@@ -453,19 +293,6 @@ public enum AgentActionJSONParser {
             }
         }
         return nil
-    }
-
-    private static func stripFences(_ text: String) -> String {
-        var output = text
-        if output.hasPrefix("```json") {
-            output.removeFirst("```json".count)
-        } else if output.hasPrefix("```") {
-            output.removeFirst("```".count)
-        }
-        if output.hasSuffix("```") {
-            output.removeLast("```".count)
-        }
-        return output.trimmingCharacters(in: .whitespacesAndNewlines)
     }
 
     private static func requiresNonEmptyArguments(_ toolName: String) -> Bool {
